refactor: split create commands of module and package into two modules
